// src/lib/generated/validators/plutus-economics/wares.ts
export const WaresRowSchema = z.object({
  artisan_profile_id: z.string().nullable(),
  cover_url: z.string().nullable(),
  created_at: z.string(),
  created_by: z.string(),
  currency: z.string(),
  description: z.string().nullable(),
  icon_emoji: z.string().nullable(),
  id: z.string(),
  media_urls: z.array(z.string()).nullable(),
  merchant_profile_id: z.string().nullable(),
  metadata: z.any().nullable(),
  name: z.string(),
  price: z.number().nullable(),
  pricing_model: z.enum(ENUM_VALUES.pricingModel),
  quantity_available: z.number().nullable(),
  quantity_sold: z.number(),
  requires_shipping: z.boolean(),
  residual_pool_percent: z.number().nullable(),
  shipping_info: z.any().nullable(),
  slug: z.string(),
  status: z.enum(ENUM_VALUES.contentStatus),
  updated_at: z.string(),
  updated_by: z.string().nullable(),
  ware_type: z.enum(ENUM_VALUES.wareType),
});